EIS: Fixing MYRIAD device issue in VI container.
The /dev/ file system must be mounted in container for making the MYRIAD
device to work properly. Additionally if a failure is observed on MYRIAD
device it is advisable to re-plug the device freshly for it to function
sanely in docker container environment. Hence updated README to reflect
the same.

Signed-off-by: Lalatendu das <lalatendu.das@intel.com>