eis-tool: SW Trigger Utility fix
Since SW Trigger Utility was conatinerized, there
was a sync issue b/w the sw trigger client and the
VI server. Sometimes the sw trigger client sends the
command even before the VI object was getting
initialized hence it was causing an error when the
command was trying to start/stop the ingestor.

To fix this we had to change the entrypoint of the
SW trigger service to "sleep infinity" and then
run the commands using "docker exec" to solve this
sync issue b/w the client and server app.

Signed-off-by: sagartp <Sagar.Tp@intel.com>
Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>